Extract logic into separate class
